Error en asignar ejercicios como componente de evaluación online (en sesión) 3153
Adjust to establish the session to the generated link.

The problem was that the session was not assigned at the time of creating the link. therefore the session id is assigned before generating the course query, this makes it possible to display the selector of the elements assigned to the session and also to the course when they are active.